% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scales.R
\name{ggplot2-scales}
\alias{ggplot2-scales}
\alias{scale_colour_palap}
\alias{scale_color_palap}
\alias{scale_fill_palap}
\alias{scale_colour_palap_d}
\alias{scale_color_palap_d}
\alias{scale_fill_palap_d}
\title{Scales to use for ggplot2}
\usage{
scale_colour_palap(..., alpha = NULL, begin = 0, end = 1,
  direction = 1, palette = "bilbao")

scale_color_palap(..., alpha = NULL, begin = 0, end = 1,
  direction = 1, palette = "bilbao")

scale_fill_palap(..., alpha = NULL, begin = 0, end = 1,
  direction = 1, palette = "bilbao")

scale_colour_palap_d(..., alpha = 1, begin = 0, end = 1,
  direction = 1, palette = "batlow", aesthetics = "colour")

scale_color_palap_d(..., alpha = 1, begin = 0, end = 1,
  direction = 1, palette = "batlow", aesthetics = "colour")

scale_fill_palap_d(..., alpha = 1, begin = 0, end = 1,
  direction = 1, palette = "batlow", aesthetics = "fill")
}
\arguments{
\item{...}{Arguments to pass on to `ggplot2::scale_colour_gradientn()`,
`ggplot2::scale_fill_gradientn()`, `ggplot2::ggplot2::discrete_scale()`}

\item{alpha}{The opacity of the generated colours. If specified rgba values
will be generated. The default (`NULL`) will generate rgb values which
corresponds to `alpha = 1`}

\item{begin}{The interval within the palette to sample colours from.
Defaults to `0` and `1` respectively}

\item{end}{The interval within the palette to sample colours from.
Defaults to `0` and `1` respectively}

\item{direction}{Either `1` or `-1`. If `-1` the palette will be reversed}

\item{palette}{The name of the palette to sample from. See
`scico::scico_palette_names` for a list of possible names}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful,
for example, to apply colour settings to the colour and fill aesthetics at
the same time, via aesthetics = c("colour", "fill").}
}
\value{
A `ScaleContinuous` or `ScaleDiscrete` object that can be added to a
  `ggplot` object
}
\description{
These functions provide the option to use the scico/palap palettes along
   with the `ggplot2` package. It goes without saying that it requires
   `ggplot2` to work.
}
\examples{

if (require('ggplot2')) {
  volcano <- data.frame(
    x = rep(seq_len(ncol(volcano)), each = nrow(volcano)),
    y = rep(seq_len(nrow(volcano)), ncol(volcano)),
    height = as.vector(volcano)
  )

  ggplot(volcano, aes(x = x, y = y, fill = height)) +
    geom_raster() +
    scale_fill_palap(palette = 'tokyo')

  ggplot(iris, aes(x=Petal.Width, y=Petal.Length)) +
    geom_point(aes(color=Species), size=10) +
    scale_colour_palap_d()
}



}
