% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palap.R
\name{palap}
\alias{palap}
\title{Symmetric colour palette}
\usage{
palap(n, alpha = NULL, begin = 0, end = 1, direction = 1,
  palette = "PuBu", package = "RColorBrewer")
}
\arguments{
\item{n}{The number of colours to generate for the palette}

\item{alpha}{The opacity of the generated colours. If specified rgba values
will be generated. The default (`NULL`) will generate rgb values which
corresponds to `alpha = 1`}

\item{begin, end}{The interval within the palette to sample colours from.
Defaults to `0` and `1` respectively}

\item{direction}{Either `1` or `-1`. If `-1` the palette will be reversed}

\item{palette}{The name of the palette to sample from from the `package`. See
`paletteer::palettes_d_names` for a list of possible names. Default is
"BuPu" from RColorBrewer}

\item{package}{The package to take the palette from}
}
\value{
A character vector of length `n` with hexencoded rgb(a) colour values
}
\description{
This is currently just wrapping `scico::scico` with some tricks
}
\examples{
library(prismatic)
plot(colour(palap(8)))
plot(colour(palap(8, direction = -1)))
}
